+++ frontend/src/common/components/AnimatedText.tsx
