+++ frontend/src/common/utils/book.utils.ts
