+++ frontend/src/common/utils/config.utils.ts
