+++ frontend/src/module/book/components/SidePanel/ImageConverter.tsx
@@ -1,7 +1,7 @@
 import React, { /* useContext,*/ useState } from 'react';
 import axios from 'axios';
-import { getAPIURL } from '@/utils/api';
-import { Tooltip } from '@/components/Tooltip';
+import { getAPIURL } from '@/common/utils/api';
+import { Tooltip } from '@components/Tooltip';
 import { InformationCircleIcon } from '@heroicons/react/24/outline';
 // import { BookContext } from '../../book.context';
