+++ frontend/src/module/book/sidePanel/Item.tsx
@@ -1,4 +1,4 @@
-import { Tooltip } from '@/components/Tooltip';
+import { Tooltip } from '@components/Tooltip';
 
 const Item: React.FC<{
   className?: string;
