+++ frontend/src/module/home/UserBook.tsx
@@ -3,8 +3,8 @@ import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { Book } from '@/types/book';
-import { getMediaUrl } from '@/utils/api';
-import { Tooltip } from '@/components/Tooltip';
+import { getMediaUrl } from '@/common/utils/api';
+import { Tooltip } from '@components/Tooltip';
 
 interface BookItemProps {
   index: number;
