+++ frontend/src/services/api.service.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { Book } from '@/types/book';
 import { GraphicAsset } from '@/types/graphic-asset.entity';
-import { getAPIURL } from '@/utils/api';
+import { getAPIURL } from '@/common/utils/api';
 
 class APIService {
   // –––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
