+++ frontend/vite.config.ts
@@ -39,7 +39,7 @@ export default defineConfig(() => {
       alias: {
         '@': path.resolve(__dirname, './src'),
         '@assets': path.resolve(__dirname, './src/assets'),
-        '@components': path.resolve(__dirname, './src/components'),
+        '@components': path.resolve(__dirname, './src/common/components'),
         '@shared': path.resolve(__dirname, './src/shared'),
       },
     },
